fix(...) : chat view 관련 파일, directory 정리
