+++ app/src/main/java/com/manna/view/chat/ChatFragment.kt
@@ -1,4 +1,4 @@
-package com.manna.view
+package com.manna.view.chat
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,131 +6,21 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.gson.Gson
 import com.manna.ChatResponse
 import com.manna.Logger
 import com.manna.R
 import com.manna.UserHolder
 import com.manna.databinding.FragmentChatBinding
-import com.manna.databinding.ItemChatBinding
-import com.manna.databinding.ItemMyChatBinding
 import com.manna.ext.HeightProvider
 import com.manna.ext.ViewUtil
 import com.manna.network.api.MeetApi
-import com.manna.view.location.MeetDetailActivity
 import io.socket.client.IO
 import io.socket.client.Manager
 import io.socket.client.Socket
 import io.socket.emitter.Emitter
-import kotlinx.android.synthetic.main.activity_meet_detail.*
 import java.net.URI
-import java.text.SimpleDateFormat
-import java.util.*
-
-
-class ChatAdapter :
-    ListAdapter<ChatResponse, ChatAdapterViewHolder>(
-        object : DiffUtil.ItemCallback<ChatResponse>() {
-            override fun areItemsTheSame(oldItem: ChatResponse, newItem: ChatResponse): Boolean =
-                oldItem.message?.createTimestamp == newItem.message?.createTimestamp
-
-
-            override fun areContentsTheSame(oldItem: ChatResponse, newItem: ChatResponse): Boolean =
-                oldItem == newItem
-
-        }) {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ChatAdapterViewHolder =
-        when (viewType) {
-            CHAT -> ChatAdapterViewHolder.ChatViewHolder(
-                LayoutInflater.from(parent.context).inflate(R.layout.item_chat, parent, false)
-            )
-            MY_CHAT -> ChatAdapterViewHolder.MyChatViewHolder(
-                LayoutInflater.from(parent.context).inflate(R.layout.item_my_chat, parent, false)
-            )
-            else -> error("Invalid viewType")
-        }
-
-
-    override fun onBindViewHolder(holder: ChatAdapterViewHolder, position: Int) {
-        when (holder) {
-            is ChatAdapterViewHolder.ChatViewHolder -> holder.bind(currentList[position])
-            is ChatAdapterViewHolder.MyChatViewHolder -> holder.bind(currentList[position])
-        }
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        return when (currentList[position].sender?.deviceToken) {
-            UserHolder.userResponse?.deviceId -> MY_CHAT
-            else -> CHAT
-        }
-    }
-
-    companion object {
-        private const val CHAT = 0
-        private const val MY_CHAT = 1
-    }
-}
-
-
-sealed class ChatAdapterViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-
-    class ChatViewHolder(view: View) : ChatAdapterViewHolder(view) {
-
-        private val binding = DataBindingUtil.bind<ItemChatBinding>(itemView)!!
-
-        fun bind(item: ChatResponse) {
-
-            Logger.d("${item.type}")
-            when (item.type) {
-                ChatResponse.Type.CHAT -> {
-                    binding.message.text = item.message?.message.orEmpty()
-                    binding.name.text = item.sender?.username.orEmpty()
-
-                    item.message?.createTimestamp?.let {
-                        binding.date.text =
-                            SimpleDateFormat(
-                                "h:mm",
-                                Locale.KOREA
-                            ).format(item.message?.createTimestamp)
-                    }
-                }
-                ChatResponse.Type.JOIN, ChatResponse.Type.LEAVE -> {
-
-                }
-            }
-        }
-    }
-
-    class MyChatViewHolder(view: View) : ChatAdapterViewHolder(view) {
-        private val binding = DataBindingUtil.bind<ItemMyChatBinding>(itemView)!!
-
-        fun bind(item: ChatResponse) {
-            when (item.type) {
-                ChatResponse.Type.CHAT -> {
-                    binding.message.text = item.message?.message.orEmpty()
-
-                    item.message?.createTimestamp?.let {
-                        binding.date.text =
-                            SimpleDateFormat(
-                                "h:mm",
-                                Locale.KOREA
-                            ).format(item.message?.createTimestamp)
-                    }
-                }
-                ChatResponse.Type.JOIN, ChatResponse.Type.LEAVE -> {
-
-                }
-            }
-
-        }
-    }
-}
 
 class ChatFragment : Fragment() {
 
@@ -138,12 +28,12 @@ class ChatFragment : Fragment() {
     private var keyboardHeight = 0
     private var chatSocket: Socket? = null
     private lateinit var chatAdapter: ChatAdapter
+
     private val onChatConnectReceiver = Emitter.Listener { args ->
         activity?.runOnUiThread {
             Logger.d("${args.map { it.toString() }}")
         }
     }
-
 
     private val onChatReceiver = Emitter.Listener { args ->
         activity?.runOnUiThread {
@@ -156,13 +46,6 @@ class ChatFragment : Fragment() {
             }
         }
     }
-
-    override fun onDestroy() {
-        chatSocket?.disconnect()
-        chatSocket?.off(CHAT_CONNECT, onChatReceiver)
-        super.onDestroy()
-    }
-
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -222,7 +105,7 @@ class ChatFragment : Fragment() {
 //
 //            }
 //        })
-
+//
 //        binding.inputChat.setOnFocusChangeListener { v, hasFocus ->
 //            if (hasFocus) {
 //                if (behavior.state != BottomSheetBehavior.STATE_EXPANDED) {
